P2P in join
