+++ tests/benchmarks/test_join.py
@@ -17,7 +17,7 @@ mem_mult = [
 
 
 @pytest.mark.parametrize("mem_mult", mem_mult)  # [0.1, 1, 10]
-def test_join_big(small_client, mem_mult):
+def test_join_big(small_client, mem_mult, shuffle):
     memory = cluster_memory(small_client)  # 76.66 GiB
 
     df1_big = timeseries_of_size(
@@ -34,11 +34,11 @@ def test_join_big(small_client, mem_mult):
     df2_big["x2"] = df2_big["x"] * 1e9
     df2_big = df2_big.astype({"x2": "int"})
 
-    dd.merge(df1_big, df2_big, on="x2", how="inner").compute()
+    dd.merge(df1_big, df2_big, on="x2", how="inner", shuffle=shuffle).compute()
 
 
 @pytest.mark.parametrize("mem_mult", mem_mult)  # [0.1, 1, 10]
-def test_join_big_small(small_client, mem_mult):
+def test_join_big_small(small_client, mem_mult, shuffle):
     memory = cluster_memory(small_client)  # 76.66 GiB
 
     df_big = timeseries_of_size(
@@ -56,4 +56,4 @@ def test_join_big_small(small_client, mem_mult):
     df_small["x2"] = df_small["x"] * 1e9
     df_small_pd = df_small.astype({"x2": "int"}).compute()
 
-    dd.merge(df_big, df_small_pd, on="x2", how="inner").compute()
+    dd.merge(df_big, df_small_pd, on="x2", how="inner", shuffle=shuffle).compute()
